backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
openai>=1.0.0
PyMuPDF>=1.24.0
sqlalchemy>=2.0.0
stripe>=7.0.0
PyYAML>=6.0
Jinja2>=3.1
